utils.py is not very useful as separate file
